Prevent FdData leaks in getNotification
The rationale for this change is a memory leak observed in a production
application using `postgresql-simple-0.6.4`. The related code is
proprietary but I will try to give the gist of the issue. If a minimal
repro is requested I could put one together.

When `getNotification` is blocking in an `async`-spawned thread, upon
cancelling and restarting the thread in a loop (reusing the same
connection) a heap profile shows an increasing amount of `FdData`
closures building up (along with a few other closures, e.g. the `TVar
Nothing` created in `threadWaitSTM`). Forgive my limited knowledge of
the GHC `EventManager`, but it seems that the registered callback made
by `threadWaitReadSTM` does not get removed from the `EventManager`
state when `waitRead` (I presume) is interrupted by the
`AsyncCancelled`.

This change somewhat mirrors how the non-STM `threadWait` handles
exceptions, and so I think it should be benign. I've tested this change
against our application and confirmed the `FdData` and related closures
are no longer hanging around.